Requirements_Generation_System/requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.8.0
pyyaml>=6.0
jinja2>=3.1.0
pydantic>=2.0.0
rich>=13.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
networkx>=3.0
matplotlib>=3.8.0
tqdm>=4.65.0

# Application Builder dependencies
gitpython>=3.1.0
PyGithub>=2.1.0

# Dashboard dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
watchfiles>=0.21.0

# Optional dependencies for enhanced features
# pandas>=2.0.0  # For data analysis
# plotly>=5.0.0  # For interactive visualizations
# reportlab>=4.0.0  # For PDF generation
# python-docx>=0.8.0  # For Word document generation
